blood-test-analyser-debug/requirements.txt
crewai==0.130.0
fastapi==0.110.3
openai
langchain-openai
python-dotenv
uvicorn
PyPDF2